Refactoring judicial booking
