+++ src/main/java/uk/gov/hmcts/reform/judicialbooking/launchdarkly/FeatureConditionEvaluation.java
@@ -1,8 +1,11 @@
 package uk.gov.hmcts.reform.judicialbooking.launchdarkly;
 
+import com.launchdarkly.sdk.LDUser;
+import com.launchdarkly.sdk.server.LDClient;
 import com.launchdarkly.shaded.org.jetbrains.annotations.NotNull;
-import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Lazy;
 import org.springframework.stereotype.Component;
 import org.springframework.web.servlet.HandlerInterceptor;
 import uk.gov.hmcts.reform.judicialbooking.controller.advice.exception.ForbiddenException;
@@ -11,33 +14,55 @@ import uk.gov.hmcts.reform.judicialbooking.util.SecurityUtils;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import java.util.HashMap;
 
 @Component
-@AllArgsConstructor
 public class FeatureConditionEvaluation implements HandlerInterceptor {
 
-    @Autowired
-    private final FeatureToggleService featureToggleService;
+    public static final String USER = "user";
+    public static final String SERVICE_NAME = "servicename";
+    public static final String GET = "GET";
+    public static final String POST = "POST";
+    public static final String DELETE = "DELETE";
 
     @Autowired
-    private final SecurityUtils securityUtils;
+    @Lazy
+    private SecurityUtils securityUtils;
+
+    @Autowired
+    private LDClient ldClient;
+
+    @Value("${launchdarkly.sdk.environment}")
+    private String environment;
+
+    @Value("${launchdarkly.sdk.user}")
+    private String userName;
+
+    private static final HashMap<String, String> getRequestMap = new HashMap<>();
+    private static final HashMap<String, String> postRequestMap = new HashMap<>();
+
+    static {
+        //Any new end point need to be placed in respective map.
+        postRequestMap.put("/am/bookings","create-bookings-api-flag");
+        postRequestMap.put("/am/bookings/query","query-bookings-api-flag");
+    }
 
     @Override
     public boolean preHandle(@NotNull HttpServletRequest request,
                              @NotNull HttpServletResponse response, @NotNull Object arg2) throws Exception {
 
-        String flagName = featureToggleService.getLaunchDarklyFlag(request);
+        String flagName = getLaunchDarklyFlag(request);
 
         if (flagName == null) {
             throw new ForbiddenException("The endpoint is not configured in Launch Darkly");
         }
 
-        if (!featureToggleService.isValidFlag(flagName)) {
+        if (!isValidFlag(flagName)) {
             throw new ResourceNotFoundException(String.format(
                     "The flag %s is not configured in Launch Darkly", flagName));
         }
 
-        boolean flagStatus = featureToggleService.isFlagEnabled(securityUtils.getServiceName(), flagName);
+        boolean flagStatus = isFlagEnabled(securityUtils.getServiceName(), flagName);
         if (!flagStatus) {
             throw new ForbiddenException(String.format("Launch Darkly flag is not enabled for the endpoint %s",
                     request.getRequestURI()));
@@ -45,4 +70,35 @@ public class FeatureConditionEvaluation implements HandlerInterceptor {
         return flagStatus;
     }
 
+    public boolean isFlagEnabled(String serviceName, String flagName) {
+        LDUser user = new LDUser.Builder(environment)
+                .firstName(userName)
+                .lastName(USER)
+                .custom(SERVICE_NAME, serviceName)
+                .build();
+
+        return ldClient.boolVariation(flagName, user, false);
+    }
+
+    public boolean isValidFlag(String flagName) {
+        return ldClient.isFlagKnown(flagName);
+    }
+
+    public String getLaunchDarklyFlag(HttpServletRequest request) {
+        String uri = request.getRequestURI();
+        switch (request.getMethod()) {
+            case GET:
+                if (getRequestMap.get(uri) != null) {
+                    return getRequestMap.get(uri);
+                }
+                break;
+            case POST:
+                if (postRequestMap.get(uri) != null) {
+                    return postRequestMap.get(uri);
+                }
+                break;
+            default:
+        }
+        return null;
+    }
 }
